Create a Video Card General Component
